Initial refactor reducer
